make sure to try and send signed tv monitoring.

+++ ECU/Src/devices/torquecontrol.c
@@ -68,9 +68,9 @@ void doVectoring(float Torque_Req, vectoradjust * adj, speedadjust * spd )
 			(int16_t)rtU.bus_Vehicle_acceleration,
 			(int16_t)rtU.bus_Vehicle_yaw_rate,
 			0);
-	CAN_Send4vals( 0x7CF, rtY.TCS_TCS_FL*NMSCALING, rtY.TCS_TCS_FR*NMSCALING, rtY.TCS_TCS_RL*NMSCALING, rtY.TCS_TCS_RR*NMSCALING);
+	CAN_Send4vals( 0x7CF, (int16_t)rtY.TCS_TCS_FL*NMSCALING, (int16_t)rtY.TCS_TCS_FR*NMSCALING, (int16_t)rtY.TCS_TCS_RL*NMSCALING, (int16_t)rtY.TCS_TCS_RR*NMSCALING);
 	CAN_Send4vals( 0x7D0, rtY.TCS_RPMmaxFL, rtY.TCS_RPMmaxFR, rtY.TCS_RPMmaxRL, rtY.TCS_RPMmaxRR );
-	CAN_Send4vals( 0x7D1, rtY.TV_TV_torqueFL*NMSCALING, rtY.TV_TV_torqueFR*NMSCALING, rtY.TV_TV_torqueRL*NMSCALING, rtY.TV_TV_torqueRR*NMSCALING );
+	CAN_Send4vals( 0x7D1, (int16_t)rtY.TV_TV_torqueFL*NMSCALING, (int16_t)rtY.TV_TV_torqueFR*NMSCALING, (int16_t)rtY.TV_TV_torqueRL*NMSCALING, (int16_t)rtY.TV_TV_torqueRR*NMSCALING );
 
 	int maxreq = 0;
